% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterdVCOV.R
\name{ClusteredVCOV}
\alias{ClusteredVCOV}
\title{Clustered VCOV matrix}
\usage{
ClusteredVCOV(model, dat, cluster)
}
\arguments{
\item{model}{the model object passed}

\item{dat}{the raw data used in the model}

\item{cluster}{a character designating the variable to cluster errors by}
}
\description{
\code{ClusteredVCOV} produced clusted VCOV matrix
}

