% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_hyp_test.R
\name{GFT}
\alias{GFT}
\title{GFT - Generalized F test}
\usage{
GFT(R, r, V, BetaHat)
}
\arguments{
\item{R}{A matrix of selected terms in restricitons}

\item{r}{A vector of restrictions such that matrix multiplication of R and BetaHat equals r}

\item{V}{The variance-covariance matrix produced by the regression object}

\item{BetaHat}{a vector of estimated coefficients from the model}
}
\value{
Ftest the F-statistic for the performed test
}
\description{
\code{GFT} runs a generalized F test on linear combinations of regression outputs
}
\examples{
data(iris)
reg <- lm(Sepal.Length ~ Sepal.Width + Petal.Length)

}

