% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormDiff.R
\name{NormDiff}
\alias{NormDiff}
\title{NormDiff}
\usage{
NormDiff(Data, BalanceVars, GroupVar, TreatVar)
}
\arguments{
\item{Data}{the data}

\item{BalanceVars}{the variables to check for balance in}

\item{GroupVar}{the groups to check for balance in}

\item{TreatVar}{the treatement we want balanced}
}
\description{
NormDiff
}

