% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RobustRegression.R
\name{RobustRegression}
\alias{RobustRegression}
\title{Runs robust regression}
\usage{
RobustRegression(Model, Data, ClusterVar)
}
\arguments{
\item{Model}{a linear regression object}

\item{Data}{the data used in the linear regression model}

\item{ClusterVar}{a character indicating variables to cluster erros by}
}
\description{
\code{RobustRegression} runs a robust regression with
heteroskedastic robust standard errors
}

