% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustRegression.R
\name{RobustRegression}
\alias{RobustRegression}
\title{Runs robust regression}
\usage{
RobustRegression(model, dat, cluster_var)
}
\arguments{
\item{model}{a linear regression object}

\item{dat}{the data used in the linear regression model}

\item{cluster_var}{a character indicating variables to cluster erros by}
}
\value{
A list object with components Model - the regression object, TidyModel,
 a cleaned up version of the model, GlanceModel - tidy summary of the model,
 AugModel -  the original data with some augmented things
}
\description{
\code{RobustRegression} runs a robust regression with
heteroskedastic robust standard errors
}

